LayoutTests/http/tests/security/shape-image-cors-redirect-error-message-logging-4-expected.txt

CONSOLE MESSAGE: Cross-origin redirection to mailto://example.com denied by Cross-Origin Resource Sharing policy: not allowed to follow a cross-origin CORS redirection with non CORS scheme
Verify the error message in console in case of CORS failing checks.
